# openapi.json
{
  "components": {
    "schemas": {
...
      "CreatePricingRuleRequest": {
        "additionalProperties": false,
        "description": "Volume tiers take `brackets`; every other kind takes a `rate`. A rule is about one product,\none category or, with neither, every product.",
        "properties": {
          "brackets": {
            "items": {
              "$ref": "#/components/schemas/BracketJson-Input"
            },
            "maxItems": 10,
            "title": "Brackets",
            "type": "array"
          },
          "category_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Id"
          },
          "customer_tier": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CustomerTier"
              },
              {
                "type": "null"
              }
            ],
            "description": "Required on customer tier discounts, refused on other kinds."
          },
          "kind": {
            "$ref": "#/components/schemas/RuleKind"
          },
          "name": {
            "examples": [
              "Fasteners: 100 boxes or more"
            ],
            "maxLength": 100,
            "title": "Name",
            "type": "string"
          },
          "product_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Product Id"
          },
          "rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "A fraction with up to 4 places: 0.125 is 12.5%.",
            "examples": [
              "0.05"
            ],
            "title": "Rate"
          },
          "valid_from": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Inclusive; defaults to now.",
            "title": "Valid From"
          },
          "valid_to": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Exclusive; required on promotions.",
            "title": "Valid To"
          }
        },
        "required": [
          "kind",
          "name"
        ],
        "title": "CreatePricingRuleRequest",
        "type": "object"
      },
...
      "CustomerTier": {
        "description": "Read by customer-tier discounts (M3).",
        "enum": [
          "standard",
          "silver",
          "gold"
        ],
        "title": "CustomerTier",
        "type": "string"
      },
...
      "PricingRuleOrder": {
        "enum": [
          "name",
          "-name",
          "created_at",
          "-created_at"
        ],
        "title": "PricingRuleOrder",
        "type": "string"
      },
      "PricingRulePage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PricingRuleResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "PricingRulePage",
        "type": "object"
      },
...
      "PricingRuleResponse": {
        "properties": {
          "brackets": {
            "description": "Volume tiers only, by minimum quantity.",
            "items": {
              "$ref": "#/components/schemas/BracketJson-Output"
            },
            "title": "Brackets",
            "type": "array"
          },
          "category_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Set when the rule is about one category.",
            "title": "Category Id"
          },
          "customer_tier": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CustomerTier"
              },
              {
                "type": "null"
              }
            ],
            "description": "Set on customer tier discounts."
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "kind": {
            "description": "The stage it feeds, or `margin_floor`. New kinds may appear; handle unknown ones.",
            "examples": [
              "volume_tier",
              "customer_tier",
              "promotion",
              "margin_floor"
            ],
            "title": "Kind",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "product_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Set when the rule is about one product.",
            "title": "Product Id"
          },
          "rate": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "The discount, or a margin floor's least margin. A fraction with up to 4 places: 0.125 is 12.5%. Null for volume tiers, whose rates are in their brackets.",
            "title": "Rate"
          },
          "valid_from": {
            "description": "Inclusive.",
            "format": "date-time",
            "title": "Valid From",
            "type": "string"
          },
          "valid_to": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Exclusive; null when open-ended.",
            "title": "Valid To"
          },
          "version": {
            "description": "Also sent as the ETag; send it back in If-Match to update.",
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "kind",
          "name",
          "product_id",
          "category_id",
          "customer_tier",
          "rate",
          "brackets",
          "valid_from",
          "valid_to",
          "is_active",
          "version"
        ],
        "title": "PricingRuleResponse",
        "type": "object"
      },
      "Problem": {
        "description": "An RFC 9457 problem: ``code`` identifies it; ``errors`` lists the invalid fields.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Detail"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/FieldError"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Errors"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Instance"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "default": "about:blank",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "title",
          "status",
          "code"
        ],
        "title": "Problem",
        "type": "object"
      },
...
      "RuleKind": {
        "description": "The discount kinds are the stages of the price waterfall, in this order (ADR-0004).",
        "enum": [
          "volume_tier",
          "customer_tier",
          "promotion",
          "margin_floor"
        ],
        "title": "RuleKind",
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "An access token from `POST /api/v1/auth/login`, or a service account's API key.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "pricewright-api",
    "version": "0.5.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/pricing-rules": {
      "get": {
        "operationId": "read_pricing_rules_api_v1_pricing_rules_get",
        "parameters": [
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/RuleKind"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "in": "query",
            "name": "product_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Product Id"
            }
          },
          {
            "in": "query",
            "name": "category_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category Id"
            }
          },
          {
            "in": "query",
            "name": "customer_tier",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/CustomerTier"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Customer Tier"
            }
          },
          {
            "description": "Omit it to list inactive ones too.",
            "in": "query",
            "name": "active",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Omit it to list inactive ones too.",
              "title": "Active"
            }
          },
          {
            "description": "`-` sorts descending.",
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/PricingRuleOrder",
              "default": "name",
              "description": "`-` sorts descending."
            }
          },
          {
            "description": "Page size, 1 to 100.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "description": "Page size, 1 to 100.",
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "`next_cursor` from the previous page, with the same filters and sort; omit it for the first page.",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "`next_cursor` from the previous page, with the same filters and sort; omit it for the first page.",
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PricingRulePage"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Missing or invalid credentials"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Reading needs `pricing:read`; changes need `pricing:manage` (managers and admins)"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List the pricing rules",
        "tags": [
          "pricing"
        ]
      },
      "post": {
        "operationId": "add_pricing_rule_api_v1_pricing_rules_post",
        "parameters": [
          {
            "description": "Makes the request safe to retry (ADR-0022): a retry with the same key and the same request gets back what the first one created, with `Idempotent-Replayed: true`, instead of creating it again. 1 to 255 visible ASCII characters without `\"` or `\\`, quoted as a Structured Field String or bare; a UUID is a good key. Keys are the caller's own and kept for 24 hours. The same key with a different request is a 422 (`idempotency_key_reused`); while the first request is still running, a 409 (`idempotency_key_in_use`).",
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Makes the request safe to retry (ADR-0022): a retry with the same key and the same request gets back what the first one created, with `Idempotent-Replayed: true`, instead of creating it again. 1 to 255 visible ASCII characters without `\"` or `\\`, quoted as a Structured Field String or bare; a UUID is a good key. Keys are the caller's own and kept for 24 hours. The same key with a different request is a 422 (`idempotency_key_reused`); while the first request is still running, a 409 (`idempotency_key_in_use`).",
              "examples": [
                "\"8e03978e-40d5-43e8-bc93-6894a57f9324\""
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePricingRuleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PricingRuleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Missing or invalid credentials"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Reading needs `pricing:read`; changes need `pricing:manage` (managers and admins)"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "A request with this Idempotency-Key is still running"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Invalid fields, a rule inconsistent with its kind, or an unknown product or category; or an Idempotency-Key that is invalid or was used for a different request"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Add a pricing rule",
        "tags": [
          "pricing"
        ]
      }
    },
    "/api/v1/pricing-rules/{rule_id}": {
      "get": {
        "operationId": "read_pricing_rule_api_v1_pricing_rules__rule_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "rule_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Rule Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PricingRuleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Missing or invalid credentials"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Reading needs `pricing:read`; changes need `pricing:manage` (managers and admins)"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "No such pricing rule in this tenant (ADR-0009)"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "One pricing rule",
        "tags": [
          "pricing"
        ]
      },
      "patch": {
        "operationId": "update_pricing_rule_api_v1_pricing_rules__rule_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "rule_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Rule Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-Match"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PricingRulePatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PricingRuleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Missing or invalid credentials"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Reading needs `pricing:read`; changes need `pricing:manage` (managers and admins)"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "No such pricing rule in this tenant (ADR-0009)"
          },
          "412": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "If-Match is stale: someone else changed the rule; reload it"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Invalid fields, a rule inconsistent with its kind, or an unknown product or category"
          },
          "428": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "If-Match is required"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Edit, end-date or deactivate a pricing rule",
        "tags": [
          "pricing"
        ]
      }
    },
...
  }
}